+++ client-side-js/getUI5Version.js
@@ -1,8 +1,8 @@
 /**
  * @returns {string} UI5 version number in string form
  */
-async function clientSide_getUI5Version() {
-    return await browser.executeAsync((done) => {
+async function clientSide_getUI5Version(browserInstance = browser) {
+    return await browserInstance.executeAsync((done) => {
         done(sap.ui.version)
     })
 }