+++ client-side-js/interactWithControl.js
@@ -1,5 +1,6 @@
-async function clientSide_interactWithControl(oOptions) {
-    return await browser.executeAsync((oOptions, done) => {
+async function clientSide_interactWithControl(oOptions, browserInstance) {
+    browserInstance = await Promise.resolve(browserInstance)
+    return await browserInstance.executeAsync((oOptions, done) => {
         window.wdi5.waitForUI5(
             window.wdi5.waitForUI5Options,
             () => {