+++ examples/ui5-js-app/e2e-test-config/wdi5-multiversion.js
@@ -3,17 +3,16 @@ const path = require("path")
 const replace = require("replace-in-file")
 const Launcher = require("@wdio/cli").default
 
-// lts version - with 1.71.19 being the exception
-// in that there seems to be stuff boken in ui5 > 1.71.19 <= 1.71.25
+// lts versions (> 1.60)
 // empty string will get the newest Version which can be a "SNAPSHOT" version
-const versions = ["", "1.71.19", "1.84.3"]
+const versions = ["", "1.71", "1.84", "1.96"]
 
 ;(async () => {
     for (const version of versions) {
         // create an index.html for bootstrapping per version
-        const targetIndex = path.resolve(__dirname, `webapp/index-${version}.html`)
-        const bootstrapSrc = `https://openui5nightly.hana.ondemand.com/${version}/resources/sap-ui-core.js`
-        fsExtra.copySync(path.resolve(__dirname, `webapp/index.html`), targetIndex)
+        const targetIndex = path.resolve(__dirname, `../webapp/index-${version}.html`)
+        const bootstrapSrc = `https://ui5.sap.com/${version}/resources/sap-ui-core.js`
+        fsExtra.copySync(path.resolve(__dirname, `../webapp/index.html`), targetIndex)
         const optionsIndex = {
             files: targetIndex,
             from: [/src=\".*\"/, /"sap_horizon"/],
@@ -30,7 +29,7 @@ const versions = ["", "1.71.19", "1.84.3"]
             from: [/url: "#"/, /specs: \[.*\]/],
             to: [
                 `url: "index-${version}"`, // this is only b/c of the "soerver" webserver in use...
-                `specs: ["${path.resolve(__dirname, "webapp/test/e2e/properties-matcher.test.js")}"]`
+                `specs: ["${path.resolve(__dirname, "../webapp/test/e2e/properties-matcher.test.js")}"]`
             ]
         }
         await replace(optionsWdioConf)