+++ examples/ui5-js-app/e2e-test-config/wdio-ui5-late.conf.js
@@ -5,7 +5,7 @@ const merge = require("deepmerge")
 const _config = {
     wdi5: {
         skipInjectUI5OnStart: true,
-        url: "/js-soft/wdi5/"
+        url: "/ui5-community/wdi5/"
     },
     specs: [join("webapp", "test", "e2e", "ui5-late.test.js")],
     baseUrl: "https://github.com"