+++ examples/ui5-js-app/e2e-test-config/wdio-webserver.conf.js
@@ -7,7 +7,7 @@ const _config = {
         url: "#"
     },
     specs: [join("webapp", "test", "e2e", "**/*.test.js")],
-    exclude: [join("webapp", "test", "e2e", "ui5-late.test.js")],
+    exclude: [join("webapp", "test", "e2e", "ui5-late.test.js"), join("webapp", "test", "e2e", "multiremote.test.js")],
     baseUrl: "http://localhost:8888"
 }