+++ src/lib/wdi5-control.ts
@@ -31,12 +31,32 @@ export class WDI5Control {
     _generatedUI5Methods: Array<string>
     _initialisation = false
     _forceSelect = false
+    _logging: boolean
     _wdioBridge = <WebdriverIO.Element>{}
     _generatedWdioMethods: Array<string>
     _domId: string
 
+    // these controls receive the specific "interaction": "press" operator
+    // instead of a regular "click" event
+    // see https://openui5.hana.ondemand.com/api/sap.ui.test.actions.Press#properties
+    _specificInteractionControls = [
+        "sap.m.ComboBox",
+        "sap.m.SearchField",
+        "sap.m.Input",
+        "sap.m.List",
+        "sap.m.Table",
+        "sap.m.ObjectIdentifier",
+        "sap.m.ObjectAttribute",
+        "sap.m.Page",
+        "sap.m.semantic.FullscreenPage",
+        "sap.m.semantic.DetailPage",
+        "sap.ui.comp.smartfilterbar.SmartFilterBar"
+    ]
+
+    _browserInstance: WebdriverIO.Browser
     constructor(oOptions) {
         const {
+            browserInstance,
             controlSelector,
             wdio_ui5_key,
             forceSelect,
@@ -50,6 +70,7 @@ export class WDI5Control {
         this._wdio_ui5_key = wdio_ui5_key
         this._forceSelect = forceSelect
         this._generatedUI5Methods = generatedUI5Methods
+        this._browserInstance = browserInstance
         this._webElement = webElement
         this._webdriverRepresentation = webdriverRepresentation
         this._domId = domId
@@ -70,12 +91,15 @@ export class WDI5Control {
         this._controlSelector = controlSelector
         this._wdio_ui5_key = controlSelector.wdio_ui5_key
         this._forceSelect = forceSelect
+        this._logging = this._controlSelector?.logging ?? true
 
         const controlResult = await this._getControl()
 
         if (controlResult.status === 1) {
             // result is string and has error text -> its an error
-            Logger.error(`error retrieving control: ${this._wdio_ui5_key}`)
+            if (this._logging) {
+                Logger.error(`error retrieving control: ${this._wdio_ui5_key}`)
+            }
             return this
         } else {
             this._webElement = controlResult.domElement
@@ -121,20 +145,16 @@ export class WDI5Control {
     }
 
     /**
+     * tries to retrieve the webdriver representation of the current wdi5 control
      * @return {WebdriverIO.Element} the webdriver Element
      */
     async getWebElement() {
-        if (!this._webdriverRepresentation) {
-            // to enable transition from wdi5 to wdio api in allControls
-            await this.renewWebElement()
-        }
-
-        if (util.types.isProxy(this.getWebElement)) {
-            const id = await Promise.resolve(this._domId)
-            const el = await $(`//*[@id="${id}"]`)
-            return el
-        } else {
-            return this._webdriverRepresentation
+        try {
+            return await this._getWebElement()
+        } catch (error) {
+            if (this._logging) {
+                Logger.error(`cannot call "getWebElement()", because ${error.message}`)
+            }
         }
     }
 
@@ -147,28 +167,18 @@ export class WDI5Control {
     }
 
     /**
-     * @param id
-     * @returns
-     */
-    async renewWebElement(id: string = this._domId) {
-        this._webdriverRepresentation = await $(`//*[@id="${id}"]`)
-        return this._webdriverRepresentation
-    }
-
-    /**
      * bridge to UI5 control api "getAggregation"
      * @param name name of the aggregation
      * @return array of UI5 controls representing the aggregation
      */
     async getAggregation(name: string) {
-        const _forceSelect: boolean = util.types.isProxy(this._forceSelect)
-            ? await Promise.resolve(this._forceSelect)
-            : this._forceSelect
-
-        if (_forceSelect) {
-            await this._renewWebElementReference()
+        try {
+            return await this._getAggregation(name)
+        } catch (error) {
+            if (this._logging) {
+                Logger.error(`cannot get aggregation "${name}", because ${error.message}`)
+            }
         }
-        return await this._getAggregation(name)
     }
 
     /**
@@ -176,17 +186,29 @@ export class WDI5Control {
      * @param text
      */
     async enterText(text: string) {
-        // if (this._forceSelect) {
-        //     this._renewWebElementReference();
-        // }
-
+        let selector
+        let logging
+        if (util.types.isProxy(this._controlSelector)) {
+            const _controlSelector = await Promise.resolve(this._controlSelector)
+            selector = await Promise.resolve(_controlSelector.selector)
+            logging = await Promise.resolve(this._logging)
+        } else {
+            selector = this._controlSelector.selector
+            logging = this._logging
+        }
         const oOptions = {
             enterText: text,
-            selector: this._controlSelector.selector,
+            selector,
             clearTextFirst: true,
             interactionType: "ENTER_TEXT"
         }
-        await this.interactWithControl(oOptions)
+        try {
+            await this._interactWithControl(oOptions)
+        } catch (error) {
+            if (logging) {
+                Logger.error(`cannot call enterText(), because ${error.message}`)
+            }
+        }
         return this
     }
 
@@ -195,29 +217,56 @@ export class WDI5Control {
      * this works both on a standalone control as well as with the fluent async api
      */
     async press() {
+        // support fluent async api
+        let className
+        let controlSelector
+        let specificInteractionControls
+        let logging
         if (util.types.isProxy(this._domId)) {
-            const id = await Promise.resolve(this._domId)
-            const webelement = await $(`//*[@id="${id}"]`)
-            await webelement.click()
+            specificInteractionControls = await Promise.resolve(this._specificInteractionControls)
+            const _controlInfo = await Promise.resolve(this._metadata)
+            className = _controlInfo.className
+            controlSelector = await Promise.resolve(this._controlSelector)
+            logging = await Promise.resolve(this._logging)
         } else {
-            await ((await this.getWebElement()) as unknown as WebdriverIO.Element).click()
+            specificInteractionControls = this._specificInteractionControls
+            className = this.getControlInfo().className
+            controlSelector = this._controlSelector
+            logging = this._logging
         }
+
+        // route operations on a sap.m.SearchField
+        // via the RecordReplay api instead of via WebdriverIO
+        if (
+            specificInteractionControls.includes(className as string) &&
+            controlSelector.selector.interaction.match(/press/i)
+        ) {
+            if (logging) {
+                Logger.info(`using OPA5 Press action to interact with this ${className}...`)
+            }
+            const oOptions = {
+                selector: controlSelector.selector,
+                interactionType: "PRESS"
+            }
+            try {
+                await this._interactWithControl(oOptions)
+            } catch (error) {
+                if (logging) {
+                    Logger.error(`cannot issue OPA5-press() on control, because ${error.message}`)
+                }
+            }
+        } else {
+            // interact via wdio
+            try {
+                await ((await this._getWebElement()) as unknown as WebdriverIO.Element).click()
+            } catch (error) {
+                if (logging) {
+                    Logger.error(`cannot call press(), because ${error.message}`)
+                }
+            }
+        }
+
         return this
-    }
-
-    /**
-     * Interact with specific control.
-     * @param {object} oOptions
-     * @param {sap.ui.test.RecordReplay.ControlSelector} oOptions.selector - UI5 type
-     * @param {sap.ui.test.RecordReplay.InteractionType} oOptions.interactionType - UI5 type
-     * @param {string} oOptions.enterText
-     * @param {boolean} oOptions.clearTextFirst
-     */
-    async interactWithControl(oOptions) {
-        const result = (await clientSide_interactWithControl(oOptions)) as clientSide_ui5Response
-
-        this._writeObjectResultLog(result, "interactWithControl()")
-        return result.result
     }
 
     /**
@@ -231,8 +280,15 @@ export class WDI5Control {
         if (oOptions?.eval) {
             oOptions = "(" + oOptions.eval.toString() + ")"
         }
-        const result = (await clientSide_fireEvent(webElement, eventName, oOptions)) as clientSide_ui5Response
-        this._writeObjectResultLog(result, "fireEvent()")
+        const result = (await clientSide_fireEvent(
+            webElement,
+            eventName,
+            oOptions,
+            this._browserInstance
+        )) as clientSide_ui5Response
+        if (this._logging) {
+            this._writeObjectResultLog(result, "fireEvent()")
+        }
         return result.result
     }
 
@@ -247,6 +303,76 @@ export class WDI5Control {
     }
 
     // --- private methods ---
+
+    /**
+     * Interact with specific control.
+     * @param {object} oOptions
+     * @param {sap.ui.test.RecordReplay.ControlSelector} oOptions.selector - UI5 type
+     * @param {sap.ui.test.RecordReplay.InteractionType} oOptions.interactionType - UI5 type
+     * @param {string} oOptions.enterText
+     * @param {boolean} oOptions.clearTextFirst
+     */
+    private async _interactWithControl(oOptions) {
+        // const domId = util.types.isProxy(this._domId) ? await Promise.resolve(this._domId) : this._domId
+        let domId
+        let logging
+        if (util.types.isProxy(this._domId)) {
+            domId = await Promise.resolve(this._domId)
+            logging = await Promise.resolve(this._logging)
+        } else {
+            domId = this._domId
+            logging = this._logging
+        }
+        const browserInstance = util.types.isProxy(this._browserInstance)
+            ? await Promise.resolve(this._browserInstance)
+            : this._browserInstance
+
+        if (domId) {
+            const result = (await clientSide_interactWithControl(oOptions, browserInstance)) as clientSide_ui5Response
+            if (logging) {
+                this._writeObjectResultLog(result, "interactWithControl()")
+            }
+            // return result.result
+            return this
+        } else {
+            throw Error("control could not be found")
+        }
+    }
+
+    /**
+     * returns the wdio web element.
+     * @throws will throw an error when no DOM Element was found
+     * @return {WebdriverIO.Element} the webdriver Element
+     */
+    private async _getWebElement() {
+        if (util.types.isProxy(this._domId)) {
+            const id = await Promise.resolve(this._domId)
+            if (id) {
+                const webElement = await $(`//*[@id="${id}"]`)
+                return webElement
+            } else {
+                throw Error("control could not be found")
+            }
+        }
+        if (!this._webdriverRepresentation) {
+            // to enable transition from wdi5 to wdio api in allControls
+            await this._renewWebElement()
+        }
+        return this._webdriverRepresentation
+    }
+
+    /**
+     * @param id
+     * @returns
+     */
+    private async _renewWebElement(id: string = this._domId) {
+        if (this._domId) {
+            this._webdriverRepresentation = await this._browserInstance.$(`//*[@id="${id}"]`)
+            return this._webdriverRepresentation
+        } else {
+            throw Error("control could not be found")
+        }
+    }
 
     /**
      * retrieve UI5 control represenation of a UI5 control's aggregation
@@ -271,12 +397,14 @@ export class WDI5Control {
                 }
 
                 // get WDI5 control
-                aResultOfPromises.push(browser.asControl(selector))
+                aResultOfPromises.push(this._browserInstance.asControl(selector))
             })
 
             return await Promise.all(aResultOfPromises)
         } else {
-            Logger.warn(`${this._wdio_ui5_key} has no aControls`)
+            if (this._logging) {
+                Logger.warn(`${this._wdio_ui5_key} has no aControls`)
+            }
         }
     }
 
@@ -301,9 +429,11 @@ export class WDI5Control {
             }
 
             // get WDI5 control
-            eResult = await browser.asControl(selector)
+            eResult = await this._browserInstance.asControl(selector)
         } else {
-            Logger.warn(`${this._wdio_ui5_key} has no aControls`)
+            if (this._logging) {
+                Logger.warn(`${this._wdio_ui5_key} has no aControls`)
+            }
         }
 
         return eResult
@@ -321,7 +451,9 @@ export class WDI5Control {
                 this[sMethodName] = this._executeControlMethod.bind(this, sMethodName, this._webElement)
             })
         } else {
-            Logger.warn(`${this._wdio_ui5_key} has no sReplFunctionNames`)
+            if (this._logging) {
+                Logger.warn(`${this._wdio_ui5_key} has no sReplFunctionNames`)
+            }
         }
     }
 
@@ -334,7 +466,9 @@ export class WDI5Control {
                 }
             })
         } else {
-            Logger.warn(`${this._wdio_ui5_key} has no sReplFunctionNames`)
+            if (this._logging) {
+                Logger.warn(`${this._wdio_ui5_key} has no sReplFunctionNames`)
+            }
         }
     }
 
@@ -351,7 +485,13 @@ export class WDI5Control {
         ...args
     ) {
         if (this._forceSelect) {
-            this._webElement = await this._renewWebElementReference()
+            try {
+                this._webElement = await this._renewWebElementReference()
+            } catch (error) {
+                if (this._logging) {
+                    Logger.error(`cannot execute ${methodName}(), because ${error.message}`)
+                }
+            }
         }
         // special case for custom data attached to a UI5 control:
         // pass the arguments to the event handler (like UI5 handles and expects them) also
@@ -364,7 +504,12 @@ export class WDI5Control {
         // returns the array of [0: "status", 1: result]
 
         // regular browser-time execution of UI5 control method
-        const result = (await clientSide_executeControlMethod(webElement, methodName, args)) as clientSide_ui5Response
+        const result = (await clientSide_executeControlMethod(
+            webElement,
+            methodName,
+            this._browserInstance,
+            args
+        )) as clientSide_ui5Response
 
         // create logging
         this._writeObjectResultLog(result, methodName)
@@ -430,18 +575,38 @@ export class WDI5Control {
      *
      * @param aggregationName
      * @param webElement
+     * @throws will throw an error when no webElement was found
      * @return {any}
      */
     private async _getAggregation(
         aggregationName: string,
         webElement: WebdriverIO.Element | string = this._webElement
     ) {
+        const _forceSelect: boolean = util.types.isProxy(this._forceSelect)
+            ? await Promise.resolve(this._forceSelect)
+            : this._forceSelect
+
+        const _logging: boolean = util.types.isProxy(this._logging)
+            ? await Promise.resolve(this._logging)
+            : this._logging
+
+        if (_forceSelect) {
+            await this._renewWebElementReference()
+        }
         if (util.types.isProxy(webElement)) {
             webElement = await Promise.resolve(webElement)
         }
-        const result = (await clientSide_getAggregation(webElement, aggregationName)) as clientSide_ui5Response
-
-        this._writeObjectResultLog(result, "_getAggregation()")
+        if (!webElement) {
+            throw Error("control could not be found")
+        }
+        const result = (await clientSide_getAggregation(
+            webElement,
+            aggregationName,
+            this._browserInstance
+        )) as clientSide_ui5Response
+        if (_logging) {
+            this._writeObjectResultLog(result, "_getAggregation()")
+        }
 
         let wdiItems = []
         if (result.status === 0) {
@@ -458,9 +623,13 @@ export class WDI5Control {
      * this method is also used wdi5-internally to implement the extended forceSelect option
      */
     private async _renewWebElementReference() {
-        const newWebElement = (await this._getControl({ selector: { id: this._domId } })).domElement // added to have a more stable retrieval experience
-        this._webElement = newWebElement
-        return newWebElement
+        if (this._domId) {
+            const newWebElement = (await this._getControl({ selector: { id: this._domId } })).domElement // added to have a more stable retrieval experience
+            this._webElement = newWebElement
+            return newWebElement
+        } else {
+            throw Error("control could not be found")
+        }
     }
 
     /**
@@ -499,7 +668,7 @@ export class WDI5Control {
             controlSelector.selector.properties.text = controlSelector.selector.properties.text.toString()
         }
 
-        const _result = (await clientSide_getControl(controlSelector)) as clientSide_ui5Response
+        const _result = (await clientSide_getControl(controlSelector, this._browserInstance)) as clientSide_ui5Response
         const { status, domElement, id, aProtoFunctions, className } = _result
 
         if (status === 0 && id) {
@@ -513,8 +682,9 @@ export class WDI5Control {
             // set the succesful init param
             this._initialisation = true
         }
-
-        this._writeObjectResultLog(_result, "_getControl()")
+        if (this._logging) {
+            this._writeObjectResultLog(_result, "_getControl()")
+        }
 
         return { status: status, domElement: domElement, aProtoFunctions: aProtoFunctions }
     }